pyproject.toml
[tool.poetry]
name = "gate-seal"
version = "0.1.0"
description = ""
authors = ["Azat Serikov <azatsdev@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
eth-ape = "^0.6.2"
vyper = "0.3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
